backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.22
boto3==1.35.0
httpx==0.27.0
